Added files via upload
Here's what I have so far for the GUI. Currently it can take 2 images from the user (that the user searched through a file finder GUI) and store them internally, ready to be passed to the resizing module. It also will pass the knowledge of if the user wants to pick the base image or not to the fusion module. I can add another option for deciding which algorithm to utilize depending on if the images were B&W/Color or Light vs Dark. Each step updates the progress bar to show the user visually something is being processed. I also want it to show a worded status update underneath the progress bar but I'm running into some issues with that.